deepen: absorb subscription-auth guards into Subscription module
Replaces 1 shallow module with private guards inside its sole consumer.

`lib/auth/subscription-auth.ts` exported six functions extracted for unit
testability. The deletion test exposed the shape:

| Export                          | Production callers |
|---------------------------------|--------------------|
| `verifyUsageRecordingPermission`| 1                  |
| `verifyUserExists`              | 1                  |
| `checkSubscriptionRateLimit`    | 1                  |
| `SubscriptionAuthError`         | 1                  |
| `verifySubscriptionOwnership`   | 0 (dead)           |
| `getAuthenticatedUserId`        | 0 (dead)           |
| `verifyActiveSubscription`      | 0 (dead)           |

The single caller is `services/filings/enhanced/subscriptionService.ts`. The
386-line test file `__tests__/lib/auth/subscription-auth.test.ts` asserted on
the guard predicates in isolation — the shallow extract-for-testability shape
ADR 0002 explicitly identifies as waste once the deepened module exists.

Dependency category: 2 (local-substitutable via mocked Prisma).

Tests deleted at the old interface:
- `__tests__/lib/auth/subscription-auth.test.ts` (386 lines, 18 test cases
  asserting on guard predicates in isolation).

Tests added at the new interface:
- None landed in this PR. An interface-level test file was drafted but could
  not load `subscriptionService.ts` due to a pre-existing broken import
  (`../../../lib/validation/subscription-validation` — that file does not
  exist on main). The broken module path is filed as a separate
  architecture-gap issue; this PR is scoped to the deepening alone.

Follows the ADR 0002 pattern (extract-for-testability removal): three exports
were already dead and have been dropped; the three live guards become private
implementation behind the Subscription module's existing recordFilingUsage /
getSubscriptionAnalytics interfaces.
