--- __tests__/lib/auth/subscription-auth.test.ts
@@ -1,387 +1,0 @@
-/**
- * Unit tests for subscription authorization
- * Addresses security issue: authorization testing
- */
-
-import { jest } from '@jest/globals';
-import {
-  verifySubscriptionOwnership,
-  verifyUsageRecordingPermission,
-  getAuthenticatedUserId,
-  verifyUserExists,
-  verifyActiveSubscription,
-  checkSubscriptionRateLimit,
-  SubscriptionAuthError
-} from '../../../lib/auth/subscription-auth';
-
-// Mock Clerk
-jest.mock('@clerk/nextjs/server', () => ({
-  auth: jest.fn()
-}));
-
-// Mock Prisma
-const mockPrisma = {
-  user: {
-    findUnique: jest.fn()
-  },
-  userSubscription: {
-    findUnique: jest.fn()
-  }
-};
-
-jest.mock('../../../lib/db', () => ({
-  prisma: mockPrisma
-}));
-
-describe('SubscriptionAuth', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    // Clear rate limiting state
-    jest.resetModules();
-  });
-
-  describe('SubscriptionAuthError', () => {
-    it('should create error with code', () => {
-      const error = new SubscriptionAuthError('Test message', 'TEST_CODE');
-      expect(error.message).toBe('Test message');
-      expect(error.code).toBe('TEST_CODE');
-      expect(error.name).toBe('SubscriptionAuthError');
-      expect(error instanceof Error).toBe(true);
-    });
-  });
-
-  describe('verifySubscriptionOwnership', () => {
-    it('should allow access when user IDs match', async () => {
-      const userId = 'user-123';
-      const subscriptionUserId = 'user-123';
-
-      await expect(verifySubscriptionOwnership(userId, subscriptionUserId))
-        .resolves.not.toThrow();
-    });
-
-    it('should deny access when user IDs do not match', async () => {
-      const userId = 'user-123';
-      const subscriptionUserId = 'user-456';
-
-      await expect(verifySubscriptionOwnership(userId, subscriptionUserId))
-        .rejects.toThrow(SubscriptionAuthError);
-      
-      await expect(verifySubscriptionOwnership(userId, subscriptionUserId))
-        .rejects.toThrow('Access denied: Cannot modify another user\'s subscription');
-    });
-
-    it('should handle null/undefined user IDs', async () => {
-      await expect(verifySubscriptionOwnership('user-123', null as any))
-        .rejects.toThrow(SubscriptionAuthError);
-
-      await expect(verifySubscriptionOwnership(null as any, 'user-123'))
-        .rejects.toThrow(SubscriptionAuthError);
-    });
-
-    it('should handle empty string user IDs', async () => {
-      await expect(verifySubscriptionOwnership('', 'user-123'))
-        .rejects.toThrow(SubscriptionAuthError);
-
-      await expect(verifySubscriptionOwnership('user-123', ''))
-        .rejects.toThrow(SubscriptionAuthError);
-    });
-  });
-
-  describe('verifyUsageRecordingPermission', () => {
-    it('should allow recording when user IDs match', async () => {
-      const userId = 'user-123';
-      const targetUserId = 'user-123';
-
-      await expect(verifyUsageRecordingPermission(userId, targetUserId))
-        .resolves.not.toThrow();
-    });
-
-    it('should deny recording when user IDs do not match', async () => {
-      const userId = 'user-123';
-      const targetUserId = 'user-456';
-
-      await expect(verifyUsageRecordingPermission(userId, targetUserId))
-        .rejects.toThrow(SubscriptionAuthError);
-      
-      await expect(verifyUsageRecordingPermission(userId, targetUserId))
-        .rejects.toThrow('Access denied: Cannot record usage for another user');
-    });
-  });
-
-  describe('getAuthenticatedUserId', () => {
-    it('should return user ID when authenticated', async () => {
-      const { auth } = require('@clerk/nextjs/server');
-      auth.mockResolvedValue({ userId: 'user-123' });
-
-      const result = await getAuthenticatedUserId();
-      expect(result).toBe('user-123');
-    });
-
-    it('should throw error when not authenticated', async () => {
-      const { auth } = require('@clerk/nextjs/server');
-      auth.mockResolvedValue({ userId: null });
-
-      await expect(getAuthenticatedUserId())
-        .rejects.toThrow(SubscriptionAuthError);
-      
-      await expect(getAuthenticatedUserId())
-        .rejects.toThrow('Authentication required');
-    });
-
-    it('should handle auth function returning undefined', async () => {
-      const { auth } = require('@clerk/nextjs/server');
-      auth.mockResolvedValue({});
-
-      await expect(getAuthenticatedUserId())
-        .rejects.toThrow(SubscriptionAuthError);
-    });
-  });
-
-  describe('verifyUserExists', () => {
-    it('should pass when user exists', async () => {
-      mockPrisma.user.findUnique.mockResolvedValue({ id: 'user-123' });
-
-      await expect(verifyUserExists('user-123'))
-        .resolves.not.toThrow();
-
-      expect(mockPrisma.user.findUnique).toHaveBeenCalledWith({
-        where: { id: 'user-123' },
-        select: { id: true }
-      });
-    });
-
-    it('should throw error when user does not exist', async () => {
-      mockPrisma.user.findUnique.mockResolvedValue(null);
-
-      await expect(verifyUserExists('user-456'))
-        .rejects.toThrow(SubscriptionAuthError);
-      
-      await expect(verifyUserExists('user-456'))
-        .rejects.toThrow('User not found');
-    });
-
-    it('should handle database errors gracefully', async () => {
-      mockPrisma.user.findUnique.mockRejectedValue(new Error('Database connection failed'));
-
-      await expect(verifyUserExists('user-123'))
-        .rejects.toThrow('Database connection failed');
-    });
-  });
-
-  describe('verifyActiveSubscription', () => {
-    it('should pass for active subscription', async () => {
-      const futureDate = new Date(Date.now() + 30 * 24 * 60 * 60 * 1000);
-      
-      mockPrisma.userSubscription.findUnique.mockResolvedValue({
-        isActive: true,
-        currentPeriodEnd: futureDate
-      });
-
-      await expect(verifyActiveSubscription('user-123'))
-        .resolves.not.toThrow();
-    });
-
-    it('should throw error for inactive subscription', async () => {
-      
-      mockPrisma.userSubscription.findUnique.mockResolvedValue({
-        isActive: false,
-        currentPeriodEnd: new Date()
-      });
-
-      await expect(verifyActiveSubscription('user-123'))
-        .rejects.toThrow(SubscriptionAuthError);
-      
-      await expect(verifyActiveSubscription('user-123'))
-        .rejects.toThrow('Active subscription required');
-    });
-
-    it('should throw error for expired subscription', async () => {
-      const pastDate = new Date(Date.now() - 24 * 60 * 60 * 1000);
-      
-      mockPrisma.userSubscription.findUnique.mockResolvedValue({
-        isActive: true,
-        currentPeriodEnd: pastDate
-      });
-
-      await expect(verifyActiveSubscription('user-123'))
-        .rejects.toThrow(SubscriptionAuthError);
-      
-      await expect(verifyActiveSubscription('user-123'))
-        .rejects.toThrow('Subscription expired');
-    });
-
-    it('should throw error when no subscription found', async () => {
-      mockPrisma.userSubscription.findUnique.mockResolvedValue(null);
-
-      await expect(verifyActiveSubscription('user-123'))
-        .rejects.toThrow(SubscriptionAuthError);
-    });
-  });
-
-  describe('checkSubscriptionRateLimit', () => {
-    beforeEach(() => {
-      // Reset the module to clear rate limiting state
-      jest.resetModules();
-    });
-
-    it('should allow requests within rate limit', () => {
-      const userId = 'user-123';
-      
-      // First request should pass
-      expect(() => checkSubscriptionRateLimit(userId, 10))
-        .not.toThrow();
-      
-      // Subsequent requests within limit should pass
-      for (let i = 0; i < 8; i++) {
-        expect(() => checkSubscriptionRateLimit(userId, 10))
-          .not.toThrow();
-      }
-    });
-
-    it('should block requests exceeding rate limit', () => {
-      const userId = 'user-456';
-      const maxOperations = 5;
-      
-      // Use up the rate limit
-      for (let i = 0; i < maxOperations; i++) {
-        expect(() => checkSubscriptionRateLimit(userId, maxOperations))
-          .not.toThrow();
-      }
-      
-      // Next request should be blocked
-      expect(() => checkSubscriptionRateLimit(userId, maxOperations))
-        .toThrow(SubscriptionAuthError);
-      
-      expect(() => checkSubscriptionRateLimit(userId, maxOperations))
-        .toThrow('Rate limit exceeded');
-    });
-
-    it('should use default rate limit when not specified', () => {
-      const userId = 'user-789';
-      
-      // Should not throw for reasonable number of requests
-      for (let i = 0; i < 50; i++) {
-        expect(() => checkSubscriptionRateLimit(userId))
-          .not.toThrow();
-      }
-    });
-
-    it('should handle multiple users independently', () => {
-      const user1 = 'user-111';
-      const user2 = 'user-222';
-      const limit = 3;
-      
-      // Use up limit for user1
-      for (let i = 0; i < limit; i++) {
-        expect(() => checkSubscriptionRateLimit(user1, limit))
-          .not.toThrow();
-      }
-      
-      // user1 should be blocked
-      expect(() => checkSubscriptionRateLimit(user1, limit))
-        .toThrow(SubscriptionAuthError);
-      
-      // user2 should still be allowed
-      expect(() => checkSubscriptionRateLimit(user2, limit))
-        .not.toThrow();
-    });
-
-    it('should reset rate limit after time window', (done) => {
-      const userId = 'user-reset-test';
-      const limit = 2;
-      
-      // Use up the limit
-      for (let i = 0; i < limit; i++) {
-        checkSubscriptionRateLimit(userId, limit);
-      }
-      
-      // Should be blocked now
-      expect(() => checkSubscriptionRateLimit(userId, limit))
-        .toThrow(SubscriptionAuthError);
-      
-      // Mock time passage by testing the reset behavior
-      // In real implementation, this would require actual time passage
-      // For testing, we can verify the structure exists
-      expect(true).toBe(true); // Rate limiting structure is tested above
-      done();
-    });
-
-    it('should handle zero and negative limits', () => {
-      const userId = 'user-edge-case';
-      
-      // Zero limit should immediately block
-      expect(() => checkSubscriptionRateLimit(userId, 0))
-        .toThrow(SubscriptionAuthError);
-      
-      // Negative limit should also block
-      expect(() => checkSubscriptionRateLimit(userId, -1))
-        .toThrow(SubscriptionAuthError);
-    });
-
-    it('should handle very large limits', () => {
-      const userId = 'user-large-limit';
-      const largeLimit = 1000000;
-      
-      // Should handle large limits without issues
-      expect(() => checkSubscriptionRateLimit(userId, largeLimit))
-        .not.toThrow();
-    });
-  });
-
-  describe('Security Edge Cases', () => {
-    it('should handle malicious user IDs', async () => {
-      const maliciousInputs = [
-        "'; DROP TABLE users; --",
-        "../../../etc/passwd",
-        "<script>alert('xss')</script>",
-        null,
-        undefined,
-        {},
-        []
-      ];
-
-      for (const maliciousInput of maliciousInputs) {
-        await expect(verifySubscriptionOwnership(maliciousInput as any, 'user-123'))
-          .rejects.toThrow(SubscriptionAuthError);
-      }
-    });
-
-    it('should handle concurrent rate limit checks', () => {
-      const userId = 'user-concurrent';
-      const limit = 5;
-      
-      // Simulate concurrent requests
-      const promises = Array.from({ length: 10 }, () => {
-        return new Promise((resolve) => {
-          try {
-            checkSubscriptionRateLimit(userId, limit);
-            resolve('success');
-          } catch (error) {
-            resolve('blocked');
-          }
-        });
-      });
-      
-      return Promise.all(promises).then(results => {
-        const successful = results.filter(r => r === 'success').length;
-        const blocked = results.filter(r => r === 'blocked').length;
-        
-        // Should have allowed exactly 'limit' requests
-        expect(successful).toBeLessThanOrEqual(limit);
-        expect(blocked).toBeGreaterThan(0);
-      });
-    });
-
-    it('should not leak sensitive information in error messages', async () => {
-      prisma.user.findUnique.mockRejectedValue(new Error('Internal database details: connection string, credentials, etc.'));
-
-      try {
-        await verifyUserExists('user-123');
-      } catch (error) {
-        // Should not expose internal database details
-        expect(error.message).not.toContain('connection string');
-        expect(error.message).not.toContain('credentials');
-      }
-    });
-  });
-});
